Home page moved to pages folder
