+++ client/src/App.js
@@ -2,7 +2,7 @@ import { Route, Routes } from "react-router-dom";
 
 import "./App.css";
 import AboutPratibhaNakshatra from "./AboutPratibhaNakshatra";
-import Home from "./Home";
+import Home from "./pages/Home";
 
 const App = () => (
 	<Routes>
