--- client/src/Home.css
@@ -1,12 +1,0 @@
-
-.message {
-	color: blue;
-}
-
-.message:hover {
-	background-color: yellow;
-}
-
-img{
-	width: 80%;
-}
