+++ client/src/pages/Home.js
@@ -1,41 +1,42 @@
-import { useEffect, useState } from "react";
+//import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
-
-import "./Home.css";
-import logo from "./logo.svg";
+import homepageImage from "../images/homepage-image.jpeg";
 
 export function Home() {
-	const [message, setMessage] = useState("Loading...");
+	// const [message, setMessage] = useState("Loading...");
 
-	useEffect(() => {
-		fetch("/api")
-			.then((res) => {
-				if (!res.ok) {
-					throw new Error(res.statusText);
-				}
-				return res.json();
-			})
-			.then((body) => {
-				setMessage(body.message);
-			})
-			.catch((err) => {
-				console.error(err);
-			});
-	}, []);
+	// useEffect(() => {
+	// 	fetch("/api")
+	// 		.then((res) => {
+	// 			if (!res.ok) {
+	// 				throw new Error(res.statusText);
+	// 			}
+	// 			return res.json();
+	// 		})
+	// 		.then((body) => {
+	// 			setMessage(body.message);
+	// 		})
+	// 		.catch((err) => {
+	// 			console.error(err);
+	// 		});
+	// }, []);
 
 	return (
 		<main role="main">
 			<div>
-				<img
-					className="logo"
-					data-qa="logo"
-					src={logo}
-					alt="Just the React logo"
-				/>
-				<h1 className="message" data-qa="message">
-					{message}
-				</h1>
-				<Link to="/about/this/site">About</Link>
+				<h1>Pratibha Nakshatra</h1>
+				<img src={homepageImage} style={{ width: "80%" }} alt="placeholder" />
+				<p>
+					Pratibha Nakshatra is the fusion of equality, co-existence, respect,
+					“Pure Soul”, positivity, fraternity and consciousness of origin that
+					will take everyone in the path of harmony, love and ultimately
+					achieving ‘Peace’. The scope of “Pratibha Nakshatra (Star Talent)” is
+					immense, in other words it can be a school of conscience, the
+					participants (Child, Young and Adult) will have a chance to explore
+					their origins to gain knowledge about their culture, attires,
+					traditions, language, arts, music, folklore etc so that they can
+					preserve, maintain and handover to their newer generation.
+				</p>
 			</div>
 		</main>
 	);
